refactor(lsp): split project file state
